Change type of replicationQuorum to Integer
This is prep work for sharding (https://review.ec.eng.vmware.com/#/c/21942), as
we're breaking it to multiple commits.

* Making NodeSelectorState.replicationQuorum an Integer, so that we can clearly
  distiguish between null and non-null (consistent with NodeSelectorState.
  replicationFactor).
* Making NodeSelectorState.replicationFactor an Integer, for consistency with
  replicationQuorum. 32-bit integers are more than enough for both.

Leaving membershipQuorum as a primitive (int) because it's never null.

+++ xenon-common/src/main/java/com/vmware/xenon/common/NodeSelectorState.java
@@ -55,10 +55,36 @@ public class NodeSelectorState extends ServiceDocument {
         return AVAILABLE.contains(calculateStatus(host, groupState));
     }
 
+    /**
+     * Path of this node selector's nodegroup
+     */
     public String nodeGroupLink;
-    public Long replicationFactor;
-    public int replicationQuorum;
+
+    /**
+     * Determines how many nodes need to receive a copy of a change.
+     * Null means: all nodes are replication targets.
+     */
+    public Integer replicationFactor;
+
+    /**
+     * Determines how many nodes need to accept a change for the change to succeed.
+     * Null means: use membershipQuorum as the replication quorum.
+     */
+    public Integer replicationQuorum;
+
+    /**
+     * Determines how many nodes need to be alive and reachable for the nodegroup
+     * to be considered 'available'
+     */
     public int membershipQuorum;
+
+    /**
+     * Nodegroup membership update time
+     */
     public long membershipUpdateTimeMicros;
+
+    /**
+     * Nodegroup status
+     */
     public Status status;
 }
